fixup! ♻️(tariff) linearize tariff migration history
